[update]DB接続関数の場所を変更
