+++ user/action/register.php
@@ -1,5 +1,5 @@
 <?php
-require '../../common/header.php';
+require '../../common/database.php';
 
 $user_name = $_POST['user_name'];
 $user_email = $_POST['user_email'];
